Update SVHN loader to exclude "nefarious" data
After 2 days of  debugging, it looks like tensorflow just doesn't like one particular file in the last bits of the train data. Strangely, when that particular set of images are fed to the network from a stand-still, there's no NaN's produced. To save time, updated script to exclude some last part of the train dataset. All stuff in extra dataset are still intact.